refactor: Clean things up a little bit
